+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Help screen listing valid options for vpl-inspect
+
 ### Changed
 - Session creation example to request a minimum API version